// framework/shopify/schema.d.ts
export type Maybe<T> = T | null;
export type Exact<T extends { [key: string]: unknown }> = {
	[K in keyof T]: T[K];
};
export type MakeOptional<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]?: Maybe<T[SubKey]> };
export type MakeMaybe<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]: Maybe<T[SubKey]> };
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
	ID: string;
	String: string;
	Boolean: boolean;
	Int: number;
	Float: number;
	/** An ISO-8601 encoded UTC date time string. Example value: `"2019-07-03T20:47:55Z"`. */
	DateTime: any;
	/** A signed decimal number, which supports arbitrary precision and is serialized as a string. Example value: `"29.99"`. */
	Decimal: any;
	/** A string containing HTML code. Example value: `"<p>Grey cotton knit sweater.</p>"`. */
	HTML: any;
	/** A monetary value string. Example value: `"100.57"`. */
	Money: any;
	/**
	 * An RFC 3986 and RFC 3987 compliant URI string.
	 *
	 * Example value: `"https://johns-apparel.myshopify.com"`.
	 *
	 */
	URL: any;
};
// ...
export type Checkout = Node & {
	__typename?: "Checkout";
	/** The gift cards used on the checkout. */
	appliedGiftCards: Array<AppliedGiftCard>;
	/**
	 * The available shipping rates for this Checkout.
	 * Should only be used when checkout `requiresShipping` is `true` and
	 * the shipping address is valid.
	 */
	availableShippingRates?: Maybe<AvailableShippingRates>;
	/** The date and time when the checkout was completed. */
	completedAt?: Maybe<Scalars["DateTime"]>;
	/** The date and time when the checkout was created. */
	createdAt: Scalars["DateTime"];
	/** The currency code for the Checkout. */
	currencyCode: CurrencyCode;
	/** A list of extra information that is added to the checkout. */
	customAttributes: Array<Attribute>;
	/**
	 * The customer associated with the checkout.
	 * @deprecated This field will always return null. If you have an authentication token for the customer, you can use the `customer` field on the query root to retrieve it.
	 */
	customer?: Maybe<Customer>;
	/** Discounts that have been applied on the checkout. */
	discountApplications: DiscountApplicationConnection;
	/** The email attached to this checkout. */
	email?: Maybe<Scalars["String"]>;
	/** Globally unique identifier. */
	id: Scalars["ID"];
	/** A list of line item objects, each one containing information about an item in the checkout. */
	lineItems: CheckoutLineItemConnection;
	/** The sum of all the prices of all the items in the checkout. Duties, taxes, shipping and discounts excluded. */
	lineItemsSubtotalPrice: MoneyV2;
	/** The note associated with the checkout. */
	note?: Maybe<Scalars["String"]>;
	/** The resulting order from a paid checkout. */
	order?: Maybe<Order>;
	/** The Order Status Page for this Checkout, null when checkout is not completed. */
	orderStatusUrl?: Maybe<Scalars["URL"]>;
	/**
	 * The amount left to be paid. This is equal to the cost of the line items, taxes and shipping minus discounts and gift cards.
	 * @deprecated Use `paymentDueV2` instead
	 */
	paymentDue: Scalars["Money"];
	/** The amount left to be paid. This is equal to the cost of the line items, duties, taxes and shipping minus discounts and gift cards. */
	paymentDueV2: MoneyV2;
	/**
	 * Whether or not the Checkout is ready and can be completed. Checkouts may
	 * have asynchronous operations that can take time to finish. If you want
	 * to complete a checkout or ensure all the fields are populated and up to
	 * date, polling is required until the value is true.
	 */
	ready: Scalars["Boolean"];
	/** States whether or not the fulfillment requires shipping. */
	requiresShipping: Scalars["Boolean"];
	/** The shipping address to where the line items will be shipped. */
	shippingAddress?: Maybe<MailingAddress>;
	/** The discounts that have been allocated onto the shipping line by discount applications. */
	shippingDiscountAllocations: Array<DiscountAllocation>;
	/** Once a shipping rate is selected by the customer it is transitioned to a `shipping_line` object. */
	shippingLine?: Maybe<ShippingRate>;
	/**
	 * Price of the checkout before shipping and taxes.
	 * @deprecated Use `subtotalPriceV2` instead
	 */
	subtotalPrice: Scalars["Money"];
	/** Price of the checkout before duties, shipping and taxes. */
	subtotalPriceV2: MoneyV2;
	/** Specifies if the Checkout is tax exempt. */
	taxExempt: Scalars["Boolean"];
	/** Specifies if taxes are included in the line item and shipping line prices. */
	taxesIncluded: Scalars["Boolean"];
	/**
	 * The sum of all the prices of all the items in the checkout, taxes and discounts included.
	 * @deprecated Use `totalPriceV2` instead
	 */
	totalPrice: Scalars["Money"];
	/** The sum of all the prices of all the items in the checkout, duties, taxes and discounts included. */
	totalPriceV2: MoneyV2;
	/**
	 * The sum of all the taxes applied to the line items and shipping lines in the checkout.
	 * @deprecated Use `totalTaxV2` instead
	 */
	totalTax: Scalars["Money"];
	/** The sum of all the taxes applied to the line items and shipping lines in the checkout. */
	totalTaxV2: MoneyV2;
	/** The date and time when the checkout was last updated. */
	updatedAt: Scalars["DateTime"];
	/** The url pointing to the checkout accessible from the web. */
	webUrl: Scalars["URL"];
};
// ...
export type CheckoutLineItemEdge = {
	__typename?: "CheckoutLineItemEdge";
	/** A cursor for use in pagination. */
	cursor: Scalars["String"];
	/** The item at the end of CheckoutLineItemEdge. */
	node: CheckoutLineItem;
};
// ...
export type CollectionConnection = {
	__typename?: "CollectionConnection";
	/** A list of edges. */
	edges: Array<CollectionEdge>;
	/** Information to aid in pagination. */
	pageInfo: PageInfo;
};
// ...
export type HasMetafields = {
	/** The metafield associated with the resource. */
	metafield?: Maybe<Metafield>;
	/** A paginated list of metafields associated with the resource. */
	metafields: MetafieldConnection;
};
// ...
export type ImageConnection = {
	__typename?: "ImageConnection";
	/** A list of edges. */
	edges: Array<ImageEdge>;
	/** Information to aid in pagination. */
	pageInfo: PageInfo;
};

/** List of supported image content types. */
export enum ImageContentType {
	/** A PNG image. */
	Png = "PNG",
	/** A JPG image. */
	Jpg = "JPG",
	/** A WEBP image. */
	Webp = "WEBP",
}

/** An auto-generated type which holds one Image and a cursor during pagination. */
export type ImageEdge = {
	__typename?: "ImageEdge";
	/** A cursor for use in pagination. */
	cursor: Scalars["String"];
	/** The item at the end of ImageEdge. */
	node: Image;
};
// ...
export type MediaConnection = {
	__typename?: "MediaConnection";
	/** A list of edges. */
	edges: Array<MediaEdge>;
	/** Information to aid in pagination. */
	pageInfo: PageInfo;
};
// ...
export type Metafield = Node & {
	__typename?: "Metafield";
	/** The date and time when the storefront metafield was created. */
	createdAt: Scalars["DateTime"];
	/** The description of a metafield. */
	description?: Maybe<Scalars["String"]>;
	/** Globally unique identifier. */
	id: Scalars["ID"];
	/** The key name for a metafield. */
	key: Scalars["String"];
	/** The namespace for a metafield. */
	namespace: Scalars["String"];
	/** The parent object that the metafield belongs to. */
	parentResource: MetafieldParentResource;
	/** The date and time when the storefront metafield was updated. */
	updatedAt: Scalars["DateTime"];
	/** The value of a metafield. */
	value: Scalars["String"];
	/** Represents the metafield value type. */
	valueType: MetafieldValueType;
};

/** An auto-generated type for paginating through multiple Metafields. */
export type MetafieldConnection = {
	__typename?: "MetafieldConnection";
	/** A list of edges. */
	edges: Array<MetafieldEdge>;
	/** Information to aid in pagination. */
	pageInfo: PageInfo;
};

/** An auto-generated type which holds one Metafield and a cursor during pagination. */
export type MetafieldEdge = {
	__typename?: "MetafieldEdge";
	/** A cursor for use in pagination. */
	cursor: Scalars["String"];
	/** The item at the end of MetafieldEdge. */
	node: Metafield;
};

/** A resource that the metafield belongs to. */
export type MetafieldParentResource = Product | ProductVariant;
// ...
export type MoneyV2 = {
	__typename?: "MoneyV2";
	/** Decimal money amount. */
	amount: Scalars["Decimal"];
	/** Currency of the money. */
	currencyCode: CurrencyCode;
};
// ...
export type Node = {
	/** Globally unique identifier. */
	id: Scalars["ID"];
};
// ...
export type Product = Node &
	HasMetafields & {
		__typename?: "Product";
		/** Indicates if at least one product variant is available for sale. */
		availableForSale: Scalars["Boolean"];
		/** List of collections a product belongs to. */
		collections: CollectionConnection;
		/** The compare at price of the product across all variants. */
		compareAtPriceRange: ProductPriceRange;
		/** The date and time when the product was created. */
		createdAt: Scalars["DateTime"];
		/** Stripped description of the product, single line with HTML tags removed. */
		description: Scalars["String"];
		/** The description of the product, complete with HTML formatting. */
		descriptionHtml: Scalars["HTML"];
		/**
		 * A human-friendly unique string for the Product automatically generated from its title.
		 * They are used by the Liquid templating language to refer to objects.
		 */
		handle: Scalars["String"];
		/** Globally unique identifier. */
		id: Scalars["ID"];
		/** List of images associated with the product. */
		images: ImageConnection;
		/** The media associated with the product. */
		media: MediaConnection;
		/** The metafield associated with the resource. */
		metafield?: Maybe<Metafield>;
		/** A paginated list of metafields associated with the resource. */
		metafields: MetafieldConnection;
		/**
		 * The online store URL for the product.
		 * A value of `null` indicates that the product is not published to the Online Store sales channel.
		 */
		onlineStoreUrl?: Maybe<Scalars["URL"]>;
		/** List of product options. */
		options: Array<ProductOption>;
		/** List of price ranges in the presentment currencies for this shop. */
		presentmentPriceRanges: ProductPriceRangeConnection;
		/** The price range. */
		priceRange: ProductPriceRange;
		/** A categorization that a product can be tagged with, commonly used for filtering and searching. */
		productType: Scalars["String"];
		/** The date and time when the product was published to the channel. */
		publishedAt: Scalars["DateTime"];
		/** The product's SEO information. */
		seo: Seo;
		/**
		 * A comma separated list of tags that have been added to the product.
		 * Additional access scope required for private apps: unauthenticated_read_product_tags.
		 */
		tags: Array<Scalars["String"]>;
		/** The product’s title. */
		title: Scalars["String"];
		/** The total quantity of inventory in stock for this Product. */
		totalInventory?: Maybe<Scalars["Int"]>;
		/**
		 * The date and time when the product was last modified.
		 * A product's `updatedAt` value can change for different reasons. For example, if an order
		 * is placed for a product that has inventory tracking set up, then the inventory adjustment
		 * is counted as an update.
		 */
		updatedAt: Scalars["DateTime"];
		/**
		 * Find a product’s variant based on its selected options.
		 * This is useful for converting a user’s selection of product options into a single matching variant.
		 * If there is not a variant for the selected options, `null` will be returned.
		 */
		variantBySelectedOptions?: Maybe<ProductVariant>;
		/** List of the product’s variants. */
		variants: ProductVariantConnection;
		/** The product’s vendor name. */
		vendor: Scalars["String"];
	};
// ...
export type ProductOption = Node & {
	__typename?: "ProductOption";
	/** Globally unique identifier. */
	id: Scalars["ID"];
	/** The product option’s name. */
	name: Scalars["String"];
	/** The corresponding value to the product option name. */
	values: Array<Scalars["String"]>;
};

/** The price range of the product. */
export type ProductPriceRange = {
	__typename?: "ProductPriceRange";
	/** The highest variant's price. */
	maxVariantPrice: MoneyV2;
	/** The lowest variant's price. */
	minVariantPrice: MoneyV2;
};

/** An auto-generated type for paginating through multiple ProductPriceRanges. */
export type ProductPriceRangeConnection = {
	__typename?: "ProductPriceRangeConnection";
	/** A list of edges. */
	edges: Array<ProductPriceRangeEdge>;
	/** Information to aid in pagination. */
	pageInfo: PageInfo;
};
// ...
export type ProductVariantConnection = {
	__typename?: "ProductVariantConnection";
	/** A list of edges. */
	edges: Array<ProductVariantEdge>;
	/** Information to aid in pagination. */
	pageInfo: PageInfo;
};
// ...
export type Seo = {
	__typename?: "SEO";
	/** The meta description. */
	description?: Maybe<Scalars["String"]>;
	/** The SEO title. */
	title?: Maybe<Scalars["String"]>;
};
// ...
export type SelectedOption = {
	__typename?: "SelectedOption";
	/** The product option’s name. */
	name: Scalars["String"];
	/** The product option’s value. */
	value: Scalars["String"];
};